+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[1.0.2](https://github.com/multiversx/mx-lite-wallet-dapp/pull/54)] - 2024-10-03
 
+- [Fixed faucet is missing on sovereign when recpatcha key is missing](https://github.com/multiversx/mx-lite-wallet-dapp/pull/55)
 - [Fixed page reload on network refresh](https://github.com/multiversx/mx-lite-wallet-dapp/pull/53)
 - [Updated extras API and NFTs endpoints](https://github.com/multiversx/mx-wallet-dapp/pull/52)
 - [Removed faucet when there is no recaptcha key](https://github.com/multiversx/mx-wallet-dapp/pull/51)